requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==9.0
environs[django]==9.3.2
django-phonenumber-field==7.0.0
phonenumbers==8.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0